cont

+++ src/app/Http/Controllers/ViewReservationsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Carbon\Carbon;
 
 class ViewReservationsController extends Controller
 {
@@ -20,6 +21,7 @@ class ViewReservationsController extends Controller
         if(!$conn){
             die('Could not connect '.mysqli_error());
         }
+        
         $sql = 'select * from ticket Where user_name = "'.$_COOKIE['user'].'";';
         $result = mysqli_query($conn,$sql);
         return View("/view_reservations",compact('result'));
@@ -42,12 +44,33 @@ class ViewReservationsController extends Controller
             die('Could not connect '.mysqli_error());
         }
 
+        $sql = 'select date_time  from matches m , ticket e where ticket_id  = ' .$input.'  and  e.match_id = m.match_id ';
+        $result = mysqli_query($conn,$sql);
+        $row  = mysqli_fetch_array($result);
+        $date1 = $row['date_time'];
+       
+        $current = Carbon::now();
+        $date2 = $current->subDays(3);
+        
+        $date1 = Carbon::parse($date1);
+
+        echo $date2 ;
+        echo $date1 ;
+        
+        $check = $date1->gte($date2);
+        dump($check);
+        if (!$check)
+        {
         $sql = 'Delete From  ticket  WHERE ticket_id  = ' .$input.';';
         mysqli_query($conn,$sql);
+        return response()->json(array('ticket_id'=> $input), 200);
+        }
+        else 
+        return response()->json(array('ticket_id'=> $input), 401);
 
         #AJAX Response
       
-      return response()->json(array('ticket_id'=> $input), 200);
+      
     }
 
 }
